requirements.txt
streamlit==1.32.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
Pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
scikit-image==0.21.0
pydensecrf==1.0  # Windows-compatible alternative to pydensecrf